==> Team 1/allan/02_load_harry_potter_books.R <==

install_and_load_books <- function(){
  
  get_harry_potter_books <- function(){
    if (packageVersion("devtools") < 1.6) {
      install.packages("devtools")
    }
    devtools::install_github("bradleyboehmke/harrypotter")
    library(harrypotter)
  }
}

install_and_load_books()
